backend/requirements.txt
torch
torchvision
ultralytics
fastapi
uvicorn
easyocr
opencv-python-headless
pillow
numpy
python-multipart
